impl console outputter class
